+++ test/e2e/ion-autocomplete.single-select.e2e.spec.js
@@ -102,7 +102,7 @@ describe('ion-autocomplete single select', function () {
                 // expect the callback value
                 element(by.css('input.ion-autocomplete-callback-model')).evaluate('callbackValueModel').then(function (callbackModelValue) {
                     expect(callbackModelValue.item.name).toEqual('test1');
-                    expect(callbackModelValue.selectedItems.length).toEqual(1);
+                    expect(callbackModelValue.selectedItems).toEqual(callbackModelValue.item);
                     expect(callbackModelValue.componentId).toEqual('comp1');
                 });
             })
@@ -167,7 +167,7 @@ describe('ion-autocomplete single select', function () {
 
                         // expect the callback value
                         element(by.css('input.ion-autocomplete-done-callback-model')).evaluate('doneButtonCallbackValueModel').then(function (callbackModelValue) {
-                            expect(callbackModelValue.selectedItems.length).toEqual(1);
+                            expect(callbackModelValue.selectedItems.name).toEqual("test1");
                             expect(callbackModelValue.componentId).toEqual('comp1');
                         });
                     });